Now using gevent as the server in development
Since for some reason Chrome decided to not store my cookies all of a sudden. Now I get Bad CSRF Token error whenever I send any forms to the server.

I figured the way to fix this was to get a domain name, in this case localhost.dev which meant that Chrome would support cookies.

However I encountered another problem with domain names. They are subjected to https check. So Chorme refused to connect my domain, which meant I had to get self-signed SSL certificates and install them.

The new server searches for key and cert in ssl folder with a regex. If it can't find one it defaults to cert.pem and key.pem.