Major Changes Read It
* MIGRATIONS BROKEN
Created new migrations folder so database needs to be redone.

* Word wrap added to main feed.

* Profile page now shows liked posts.
It was a major annoyance since I used "id" instead of "user_id" like in the endpoint function. It is common knowledge that Python has the word "id" reserved that's why I was getting unexpected errors.
Come to think of it errors were hinting at passing function as a value but they felt subtle at the time.

* Profile page now also shows comments of the user.