Final Commit for a while
Migrations are broken. (AGAIN)

* Fixed an issue with database and forms havng different max lengths

* Added Settings page
Now you can change email, username or password by providing the older password.

** IMPORTANT **
I am sure this may (definitely will) lead to exploits like changing your username to someone else's, or e-mail for that matter.

However this is a school project with extreme time constraints, when we include 3 week-long midterms. So it will have to do.